refactor(_common): rename file to follow convention
