+++ src/modules/_common/rate-limit/rateLimit.guard.ts
@@ -1,6 +1,6 @@
 import { ExecutionContext, Injectable } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
-import { RATE_LIMIT_METADATA } from "../const/envKeys.const";
+import { RATE_LIMIT_METADATA } from "../const/envKeys";
 import { ServiceException } from "../filter/exception/service/serviceException";
 import { RateLimitService } from "./rateLimit.service.js";
 import { Request, Response } from "express";
