+++ src/modules/auth/auth.controller.ts
@@ -19,7 +19,7 @@ import { QueryRunner } from "typeorm";
 import {
 	ENV_EMAIL_TEST_ADDRESS,
 	FRONT_ROUTE_USER_EMAIL_VERIFICATION,
-} from "../_common/const/envKeys.const";
+} from "../_common/const/envKeys";
 import { ServiceException } from "../_common/filter/exception/service/serviceException";
 import { DBQueryRunner } from "../db/query-runner/decorator/queryRunner";
 import { QueryRunnerInterceptor } from "../db/query-runner/queryRunner.interceptor";
