+++ src/modules/logger/logger.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, Logger } from "@nestjs/common";
 import * as assert from "assert";
 import { ClsService } from "nestjs-cls";
-import { TYPE_DEFAULT_VALUE } from "../_common/const/default.value";
-import { NODE_ENV } from "../_common/const/envKeys.const";
+import { TYPE_DEFAULT_VALUE } from "../_common/const/defaultValue";
+import { NODE_ENV } from "../_common/const/envKeys";
 
 const isProduction = process.env[NODE_ENV] === "production";
