+++ src/modules/painting/dto/request/updateWikiArtInfoDTO.ts
@@ -1,5 +1,5 @@
 import { PartialType } from "@nestjs/mapped-types";
-import { TYPE_DEFAULT_VALUE } from "../../../_common/const/default.value";
+import { TYPE_DEFAULT_VALUE } from "../../../_common/const/defaultValue";
 import { WikiArtPainting } from "../../entities/wikiArtPainting.entity";
 
 export class UpdateWikiArtInfoDTO extends PartialType(WikiArtPainting) {
