+++ src/modules/painting/painting.controller.ts
@@ -17,12 +17,8 @@ import {
 	ValidationPipe,
 } from "@nestjs/common";
 import { QueryRunner } from "typeorm";
-import { CONFIG_FILE_PATH } from "../_common/const/default.value";
-import {
-	AWS_BUCKET,
-	AWS_BUCKET_ARTWORK,
-	AWS_INIT_FILE_KEY_PREFIX,
-} from "../_common/const/envKeys.const";
+import { CONFIG_FILE_PATH } from "../_common/const/defaultValue";
+import { AWS_BUCKET, AWS_BUCKET_ARTWORK, AWS_INIT_FILE_KEY_PREFIX } from "../_common/const/envKeys";
 import { ServiceException } from "../_common/filter/exception/service/serviceException";
 import { S3Service } from "../aws/s3.service";
 import { DBQueryRunner } from "../db/query-runner/decorator/queryRunner";
